fix: moved TraceFlags from commands to helpers
