+++ packages/salesforcedx-utils-vscode/src/index.ts
@@ -20,7 +20,7 @@ export {
   LibraryCommandletExecutor
 } from './commands/commandletExecutors';
 export { SfdxCommandlet } from './commands/sfdxCommandlet';
-export { TraceFlags } from './commands/traceFlags';
+export { TraceFlags } from './helpers/traceFlags';
 export { ConfigSource, ConfigUtil } from './config/configUtil';
 export { OrgInfo, WorkspaceContextUtil, getLogDirPath } from './context/workspaceContextUtil';
 export {
