+++ packages/salesforcedx-utils-vscode/test/unit/helpers/traceFlags.test.ts
@@ -12,7 +12,7 @@ import { expect } from 'chai';
 import * as proxyquire from 'proxyquire';
 import { createSandbox, SinonSandbox, SinonStub } from 'sinon';
 import { nls } from '../../../src/messages';
-import { vscodeStub } from './mocks';
+import { vscodeStub } from '../commands/mocks';
 
 const { TraceFlags } = proxyquire.noCallThru()(
   '../../../src/commands',
