+++ packages/salesforcedx-vscode-apex-replay-debugger/src/commands/quickLaunch.ts
@@ -19,8 +19,8 @@ import {
   getRootWorkspacePath,
   LibraryCommandletExecutor
 } from '@salesforce/salesforcedx-utils-vscode/out/src';
-import { notificationService, TraceFlags } from '@salesforce/salesforcedx-utils-vscode/out/src/commands';
-import { getTestResultsFolder } from '@salesforce/salesforcedx-utils-vscode/out/src/helpers';
+import { notificationService } from '@salesforce/salesforcedx-utils-vscode/out/src/commands';
+import { getTestResultsFolder, TraceFlags } from '@salesforce/salesforcedx-utils-vscode/out/src/helpers';
 import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import * as path from 'path';
 import { workspace } from 'vscode';
