+++ packages/salesforcedx-vscode-apex/test/vscode-integration/commands/forceApexExecute.test.ts
@@ -8,7 +8,8 @@ import { ExecuteService } from '@salesforce/apex-node';
 import { AuthInfo, ConfigAggregator, Connection } from '@salesforce/core';
 import { MockTestOrgData, testSetup } from '@salesforce/core/lib/testSetup';
 import { getRootWorkspacePath } from '@salesforce/salesforcedx-utils-vscode/out/src';
-import { ChannelService, TraceFlags } from '@salesforce/salesforcedx-utils-vscode/out/src/commands';
+import { ChannelService } from '@salesforce/salesforcedx-utils-vscode/out/src/commands';
+import { TraceFlags } from '@salesforce/salesforcedx-utils-vscode/out/src/helpers';
 import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import { expect } from 'chai';
 import * as fs from 'fs';
